adiciona router para cliente
